% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAggregatedDrugBank.R
\name{aggregatedPlotDrugBank}
\alias{aggregatedPlotDrugBank}
\title{Create aggregated plot with mean and standard deviation of all similarity metrics against DrugBank}
\usage{
aggregatedPlotDrugBank(cosinedrugbank, dicedrugbank, jaccarddrugbank)
}
\arguments{
\item{cosinedrugbank}{data.frame containing the vector with cosine metrics for MeSH, EpSO, ESSO, and EPILONT against DrugBank}

\item{dicedrugbank}{data.frame containing the vector with dice metrics for MeSH, EpSO, ESSO, and EPILONT against DrugBank}

\item{jaccarddrugbank}{data.frame containing the vector with jaccard metrics for MeSH, EpSO, ESSO, and EPILONT against DrugBank}
}
\value{
aggdrug the plot with the aggregated graphs
}
\description{
Create aggregated plot with mean and standard deviation of all similarity metrics against DrugBank
}
\examples{
\dontrun{
aggdrug <- aggregatedPlotDrugBank(cosinedrugbank, dicedrugbank, jaccarddrugbank)
}
}
