% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAggregatedEpilepsy.R
\name{aggregatedPlotEpilepsie}
\alias{aggregatedPlotEpilepsie}
\title{Create a plot object that aggregates cosine, dice, and jaccard similarity coefficients for all three 
epilepsy ontologies EpSO, ESSO, and EPILONT}
\usage{
aggregatedPlotEpilepsie(cosineepso, cosineesso, diceepso, diceesso,
  jaccardesso, jaccardepso)
}
\arguments{
\item{cosineepso}{vector with cosine similarity coefficient values for EpSO}

\item{cosineesso}{vector with cosine similarity coefficient values for ESSO}

\item{diceepso}{vector with dice similarity coefficient values for EpSO}

\item{diceesso}{vector with dice similarity coefficient values for ESSO}

\item{jaccardesso}{vector with jaccard similarity coefficient values for ESSO}

\item{jaccardepso}{vector with jaccard similarity coefficient values for EpSO}
}
\value{
aggeps
}
\description{
Create a plot object that aggregates cosine, dice, and jaccard similarity coefficients for all three 
epilepsy ontologies EpSO, ESSO, and EPILONT
}
\examples{
\dontrun{
aggeps <- aggregatedPlotEpilepsie(cosineepso, cosineesso, \\
diceepso, diceesso, jaccardesso, jaccardepso)
}
}
