% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAggregatedMeSH.R
\name{aggregatedPlotMeSH}
\alias{aggregatedPlotMeSH}
\title{Create aggregated plot with mean and standard deviation of all similarity metrics against MeSH}
\usage{
aggregatedPlotMeSH(cosinemesh, dicemesh, jaccardmesh)
}
\arguments{
\item{cosinemesh}{data.frame containing the vector with cosine metrics for DrugBank, EpSO, ESSO, and EPILONT against MeSH}

\item{dicemesh}{data.frame containing the vector with dice metrics for DrugBank, EpSO, ESSO, and EPILONT against MeSH}

\item{jaccardmesh}{data.frame containing the vector with jaccard metrics for DrugBank, EpSO, ESSO, and EPILONT against MeSH}
}
\value{
aggmesh the plot with the aggregated graphs
}
\description{
Create aggregated plot with mean and standard deviation of all similarity metrics against MeSH
}
\examples{
\dontrun{
aggmesh <- aggregatedPlotMeSH(cosinemesh, dicemesh, jaccardmesh)
}
}
