% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSimilarities.R
\name{calcCosine}
\alias{calcCosine}
\title{Calculate the cosine similarity metric for two lists a and b}
\usage{
calcCosine(a, b)
}
\arguments{
\item{a}{list with elements that should be of same type as in list b}

\item{b}{list with elements}
}
\value{
co list with length of set b containing the cosine similarity coefficient at each position
}
\description{
Calculate the cosine similarity metric for two lists a and b
}
\examples{
calcCosine(c(1,2), c(2,3))
}
