% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstDrugBank.R
\name{calcCosineAgainstDrugBank}
\alias{calcCosineAgainstDrugBank}
\title{Calculate cosine similarity coefficients for each of the frequency vector mesh, epso
esso, and epi against the drugbank vector}
\usage{
calcCosineAgainstDrugBank()
}
\value{
dcframe dataframe with the columns for Elements, MeSH, EpSO, ESSO and EPILONT
}
\description{
Calculate cosine similarity coefficients for each of the frequency vector mesh, epso
esso, and epi against the drugbank vector
}
\examples{
dcframe <- calcCosineAgainstDrugBank ()

}
