% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcCosineAgainstMeSH}
\alias{calcCosineAgainstMeSH}
\title{Calculate cosine similarity coefficients for each of the frequency vector drugbank, epso
esso, and epi against the mesh vector}
\usage{
calcCosineAgainstMeSH()
}
\value{
mcframe dataframe with the columns for Elements, DrugBank, EpSO, ESSO and EPILONT
}
\description{
Calculate cosine similarity coefficients for each of the frequency vector drugbank, epso
esso, and epi against the mesh vector
}
\examples{
mcframe <- calcCosineAgainstMeSH ()

}
