% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcCosineDrugBankMeSH}
\alias{calcCosineDrugBankMeSH}
\title{Calculate cosine coefficient of drugbank against mesh}
\usage{
calcCosineDrugBankMeSH(drugbank, mesh)
}
\arguments{
\item{drugbank}{list with drugbank terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
cdrugbankmesh list with cosine coefficients
}
\description{
Calculate cosine coefficient of drugbank against mesh
}
\examples{
\dontrun{
calcCosineDrugBankMeSH(drugbank, mesh)
}
}
