% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcCosineEPIMeSH}
\alias{calcCosineEPIMeSH}
\title{Calculate cosine coefficient of epi against mesh}
\usage{
calcCosineEPIMeSH(epi, mesh)
}
\arguments{
\item{epi}{list with epi terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
cepimesh list with cosine coefficients
}
\description{
Calculate cosine coefficient of epi against mesh
}
\examples{
\dontrun{
calcCosineEPIMeSH(epi, mesh)
}
}
