% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcCosineEpSOMeSH}
\alias{calcCosineEpSOMeSH}
\title{Calculate cosine coefficient of epso against mesh}
\usage{
calcCosineEpSOMeSH(epso, mesh)
}
\arguments{
\item{epso}{list with epso terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
cepsomesh list with cosine coefficients
}
\description{
Calculate cosine coefficient of epso against mesh
}
\examples{
\dontrun{
calcCosineEpSOMeSH(epso, mesh)
}
}
