% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcCosineESSOMeSH}
\alias{calcCosineESSOMeSH}
\title{Calculate cosine coefficient of esso against mesh}
\usage{
calcCosineESSOMeSH(esso, mesh)
}
\arguments{
\item{esso}{list with esso terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
cepsomesh list with cosine coefficients
}
\description{
Calculate cosine coefficient of esso against mesh
}
\examples{
\dontrun{
calcCosineESSOMeSH(esso, mesh)
}
}
