% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSimilarities.R
\name{calcDice}
\alias{calcDice}
\title{Calculate the dice similarity metric for two lists a and b}
\usage{
calcDice(a, b)
}
\arguments{
\item{a}{list with elements that should be of same type as in list b}

\item{b}{list with elements}
}
\value{
di list with length of set b containing the dice similarity coefficient at each list element
}
\description{
Calculate the dice similarity metric for two lists a and b
}
\examples{
calcDice(c(1,2), c(2,3))
}
