% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstDrugBank.R
\name{calcDiceAgainstDrugBank}
\alias{calcDiceAgainstDrugBank}
\title{Calculate dice similarity coefficients for each of the frequency vector mesh, epso
esso, and epi against the drugbank vector}
\usage{
calcDiceAgainstDrugBank()
}
\value{
ddframe dataframe with the columns for Elements, MeSH, EpSO, ESSO and EPILONT
}
\description{
Calculate dice similarity coefficients for each of the frequency vector mesh, epso
esso, and epi against the drugbank vector
}
\examples{
ddframe <- calcDiceAgainstDrugBank ()

}
