% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcDiceAgainstMeSH}
\alias{calcDiceAgainstMeSH}
\title{Calculate dice similarity coefficients for each of the frequency vector drugbank, epso
esso, and epi against the mesh vector}
\usage{
calcDiceAgainstMeSH()
}
\value{
mdframe dataframe with the columns for Elements, DrugBank, EpSO, ESSO and EPILONT
}
\description{
Calculate dice similarity coefficients for each of the frequency vector drugbank, epso
esso, and epi against the mesh vector
}
\examples{
mdframe <- calcDiceAgainstMeSH ()

}
