% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcDiceDrugBankMeSH}
\alias{calcDiceDrugBankMeSH}
\title{Calculate dice coefficient of drugbank against mesh}
\usage{
calcDiceDrugBankMeSH(drugbank, mesh)
}
\arguments{
\item{drugbank}{list with drugbank terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
ddrugbankmesh list with dice coefficients
}
\description{
Calculate dice coefficient of drugbank against mesh
}
\examples{
\dontrun{
calcDiceDrugBankMeSH(drugbank, mesh)
}
}
