% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcDiceEPIMeSH}
\alias{calcDiceEPIMeSH}
\title{Calculate dice coefficient of epi against mesh}
\usage{
calcDiceEPIMeSH(epi, mesh)
}
\arguments{
\item{epi}{list with epi terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
depimesh list with dice coefficients
}
\description{
Calculate dice coefficient of epi against mesh
}
\examples{
\dontrun{
calcDiceEPIMeSH(epi, mesh)
}
}
