% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcDiceEpSOMeSH}
\alias{calcDiceEpSOMeSH}
\title{Calculate dice coefficient of epso against mesh}
\usage{
calcDiceEpSOMeSH(epso, mesh)
}
\arguments{
\item{epso}{list with epso terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
depsomesh list with dice coefficients
}
\description{
Calculate dice coefficient of epso against mesh
}
\examples{
\dontrun{
calcDiceEpSOMeSH(epso, mesh)
}
}
