% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcDiceESSOMeSH}
\alias{calcDiceESSOMeSH}
\title{Calculate dice coefficient of esso against mesh}
\usage{
calcDiceESSOMeSH(esso, mesh)
}
\arguments{
\item{esso}{list with esso terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
depsomesh list with dice coefficients
}
\description{
Calculate dice coefficient of esso against mesh
}
\examples{
\dontrun{
calcDiceESSOMeSH(esso, mesh)
}
}
