% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnrichment.R
\name{calcDSEA}
\alias{calcDSEA}
\title{Calculate dsea scores of one list in comparison to reference list}
\usage{
calcDSEA(alist, N)
}
\arguments{
\item{alist}{list of drug names to be used for calculating dsea}

\item{N}{numeric value with maximum length of lists for dsea calculation}
}
\value{
list with dsea scores
}
\description{
Calculate dsea scores of one list in comparison to reference list
}
\examples{
calcDSEA(c("Valproic acid", "Lamotrigine", "Ketamin"), 3)
}
