% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnrichment.R
\name{calcEnrichment}
\alias{calcEnrichment}
\title{Calculate enrichment of one list in comparison to reference list}
\usage{
calcEnrichment(alist, reflist)
}
\arguments{
\item{alist}{the list to compare}

\item{reflist}{the reference list}
}
\value{
list with calculated enrichment used for plotting
}
\description{
Calculate enrichment of one list in comparison to reference list
}
\examples{
a <- calcEnrichment(c("a","b","c", "d", "e"), c("b", "d"))
}
