% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSimilarities.R
\name{calcJaccard}
\alias{calcJaccard}
\title{Calculate the jaccard coefficient for two lists a and b}
\usage{
calcJaccard(a, b)
}
\arguments{
\item{a}{list with elements that should be of same type as in list b}

\item{b}{list with elements}
}
\value{
ja list with length of set b containing the jaccard similarity coefficient for each list element
}
\description{
Calculate the jaccard coefficient for two lists a and b
}
\examples{
calcJaccard(c(1,2), c(2,3))
}
