% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAggregates.R
\name{calcJaccardAgainstDrugBank}
\alias{calcJaccardAgainstDrugBank}
\title{Calculate jaccard similarity coefficients for each of the frequency vector mesh, epso
esso, and epi against the drugbank vector}
\usage{
calcJaccardAgainstDrugBank()
}
\value{
ddframe dataframe with the columns for Elements, MeSH, EpSO, ESSO and EPILONT
}
\description{
Calculate jaccard similarity coefficients for each of the frequency vector mesh, epso
esso, and epi against the drugbank vector
}
\examples{
djframe <- calcJaccardAgainstDrugBank ()
}
