% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcJaccardAgainstMeSH}
\alias{calcJaccardAgainstMeSH}
\title{Calculate jaccard similarity coefficients for each of the frequency vector drugbank, epso
esso, and epi against the mesh vector}
\usage{
calcJaccardAgainstMeSH()
}
\value{
mjframe dataframe with the columns for Elements, DrugBank, EpSO, ESSO and EPILONT
}
\description{
Calculate jaccard similarity coefficients for each of the frequency vector drugbank, epso
esso, and epi against the mesh vector
}
\examples{
mjframe <- calcJaccardAgainstMeSH ()

}
