% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcJaccardDrugBankMeSH}
\alias{calcJaccardDrugBankMeSH}
\title{Calculate jaccard coefficient of drugbank against mesh}
\usage{
calcJaccardDrugBankMeSH(drugbank, mesh)
}
\arguments{
\item{drugbank}{list with drugbank terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
jdrugbankmesh list with jaccard coefficients
}
\description{
Calculate jaccard coefficient of drugbank against mesh
}
\examples{
\dontrun{
calcJaccardDrugBankMeSH(drugbank, mesh)
}
}
