% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcJaccardEPIMeSH}
\alias{calcJaccardEPIMeSH}
\title{Calculate jaccard coefficient of epi against mesh}
\usage{
calcJaccardEPIMeSH(epi, mesh)
}
\arguments{
\item{epi}{list with epi terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
jepimesh list with jaccard coefficients
}
\description{
Calculate jaccard coefficient of epi against mesh
}
\examples{
\dontrun{
calcJaccardEPIMeSH(epi, mesh)
}
}
