% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcJaccardEpSOMeSH}
\alias{calcJaccardEpSOMeSH}
\title{Calculate jaccard coefficient of epso against mesh}
\usage{
calcJaccardEpSOMeSH(epso, mesh)
}
\arguments{
\item{epso}{list with epso terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
jepsomesh list with jaccard coefficients
}
\description{
Calculate jaccard coefficient of epso against mesh
}
\examples{
\dontrun{
calcJaccardEpSOMeSH(epso, mesh)
}
}
