% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{calcJaccardESSOMeSH}
\alias{calcJaccardESSOMeSH}
\title{Calculate jaccard coefficient of esso against mesh}
\usage{
calcJaccardESSOMeSH(esso, mesh)
}
\arguments{
\item{esso}{list with esso terms sorted by frequency}

\item{mesh}{list with mesh terms sorted by frequency}
}
\value{
jepsomesh list with jaccard coefficients
}
\description{
Calculate jaccard coefficient of esso against mesh
}
\examples{
\dontrun{
calcJaccardESSOMeSH(esso, mesh)
}
}
