% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processATC.R
\name{countATC}
\alias{countATC}
\title{Title}
\usage{
countATC(druglist, atchashda, atchashaa, atchashsec)
}
\arguments{
\item{druglist}{}

\item{atchashda}{}

\item{atchashaa}{}

\item{atchashsec}{}
}
\value{
atccounter
}
\description{
Title
}
\examples{
\dontrun{
atccounter <- countATC(druglist, atchashda, atchashaa, atchashsec)
}
}
