% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAggregates.R
\name{create_metrics}
\alias{create_metrics}
\title{Calculate required metrics for the aggregated plots against MeSH, DrugBank, and all Epilepsy Ontologies}
\usage{
create_metrics(cosine, dice, jaccard, elements)
}
\arguments{
\item{cosine}{data.frame containing the cosine vectors for MeSH, DrugBank, EpSO, ESSO, and EPILONT}

\item{dice}{data.frame containing the dice vectors for MeSH, DrugBank, EpSO, ESSO, and EPILONT}

\item{jaccard}{data.frame containing the jaccard vectors for MeSH, DrugBank, EpSO, ESSO, and EPILONT}

\item{elements}{vector of maximum length for plotting the x axis}
}
\value{
metrics data.frame containing cosine, dice, jaccard, elements, and mean and median with standard deviation
}
\description{
Calculate required metrics for the aggregated plots against MeSH, DrugBank, and all Epilepsy Ontologies
}
\examples{
dcframe <- calcCosineAgainstDrugBank ()
ddframe <- calcDiceAgainstDrugBank ()
djframe <- calcJaccardAgainstDrugBank()
meshmetrics <- create_metrics (dcframe$MeSH, ddframe$MeSH, djframe$MeSH, djframe$Elements)

}
