% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAggregates.R
\name{create_stats}
\alias{create_stats}
\title{Calculate data.frame containing stats for mean, median, and standard deviation}
\usage{
create_stats(metrics)
}
\arguments{
\item{metrics}{the data.frame return from create_metrics}
}
\value{
stats the data.frame metrics with added columns for mean, media, and standard deviation
}
\description{
Calculate data.frame containing stats for mean, median, and standard deviation
}
\examples{
\dontrun{
stats <- create_stats(metrics)
}
}
