% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmapVis.R
\name{createATCBarChart}
\alias{createATCBarChart}
\title{Creates a plot object for a given sorted hashmap.}
\usage{
createATCBarChart(shm, plottitle)
}
\arguments{
\item{shm}{hashmap that is sorted and about to be visualized}

\item{plottitle}{character vector that is the title of the plot}
}
\value{
plotbar a ggplot2 object containing the plot
}
\description{
Creates a plot object for a given sorted hashmap.
}
\examples{
\dontrun{
createATCBarChart(flmesh, "These are the most frequent ATC classes for the MeSH vector")
}
}
