% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstMeSH.R
\name{createCosineFrame}
\alias{createCosineFrame}
\title{Creates data frame for plotting the cosine coefficients against MeSH}
\usage{
createCosineFrame()
}
\value{
dcosine the data frame for the cosine coefficient that can be used by createCosinePlot
}
\description{
Creates data frame for plotting the cosine coefficients against MeSH
}
\examples{
\dontrun{
createCosineFrame()
}
}
