% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstDrugBank.R
\name{createCosineFrameDrugBank}
\alias{createCosineFrameDrugBank}
\title{Creates data frame for plotting the cosine coefficients against DrugBank}
\usage{
createCosineFrameDrugBank(mesh, drugbank, epso, esso, epi)
}
\value{
dcosinedrugbank the data frame for the cosine coefficient against DrugBank that can be used by createCosinePlotDrugBank
}
\description{
Creates data frame for plotting the cosine coefficients against DrugBank
}
\examples{
\dontrun{
createCosineFrameDrugBank(mesh, drugbank, epso, esso, epi)
}
}
