% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsEpilepsyOntologies.R
\name{createCosineFrameEpSO}
\alias{createCosineFrameEpSO}
\title{Creates data frame for plotting the cosine coefficients against EpSO}
\usage{
createCosineFrameEpSO()
}
\value{
cosineepso the data frame for the cosine coefficient against EpSO
}
\description{
Creates data frame for plotting the cosine coefficients against EpSO
}
\examples{
cosineepso <- createCosineFrameEpSO()

}
