% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsEpilepsyOntologies.R
\name{createCosineFrameESSO}
\alias{createCosineFrameESSO}
\title{Creates data frame for plotting the cosine coefficients against ESSO}
\usage{
createCosineFrameESSO(epso, esso, epi)
}
\arguments{
\item{epso}{list with epso terms sorted by frequency}

\item{esso}{list with esso terms sorted by frequency}

\item{epi}{list with epi terms sorted by frequency}
}
\value{
cosineesso the data frame for the cosine coefficient against ESSO
}
\description{
Creates data frame for plotting the cosine coefficients against ESSO
}
\examples{
\dontrun{
cosineesso <- createCosineFrameESSO(epso, esso, epi)
}
}
