% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{createCosineFrameMeSH}
\alias{createCosineFrameMeSH}
\title{Creates data frame for plotting the cosine coefficients against MeSH}
\usage{
createCosineFrameMeSH(mesh, drugbank, epso, esso, epi)
}
\value{
dcosinemesh the data frame for the cosine coefficient against MeSH that can be used by createCosinePlotMeSH
}
\description{
Creates data frame for plotting the cosine coefficients against MeSH
}
\examples{
\dontrun{
createCosineFrameMeSH()
}
}
