% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstMeSH.R
\name{createCosinePlot}
\alias{createCosinePlot}
\title{Creates the plot for all cosine coefficients against mesh}
\usage{
createCosinePlot(dcosine)
}
\arguments{
\item{dcosine}{the data frame containing the columns with the cosine coefficients}
}
\value{
cosineplot the ggplot object
}
\description{
Creates the plot for all cosine coefficients against mesh
}
\examples{
\dontrun{
createCosinePlot(createCosineFrame)
}
}
