% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstAllEpilepsy.R,
%   R/doThePlotAgainstBaseline_DB-MeSH.R
\name{createCosinePlotAll}
\alias{createCosinePlotAll}
\alias{createCosinePlotAll}
\title{Creates the plot for all cosine coefficients against drugbank}
\usage{
createCosinePlotAll(cosineepso, cosineesso)

createCosinePlotAll(cosineepso, cosineesso)
}
\arguments{
\item{cosineepso}{the data frame containing the columns with the cosine coefficients of epi and esso against epso}

\item{cosineesso}{the data frame containing the columns with the cosine coefficients of epi and epso against esso}

\item{cosineepso}{the data frame containing the columns with the cosine coefficients of epi and esso against epso}

\item{cosineesso}{the data frame containing the columns with the cosine coefficients of epi and epso against esso}
}
\value{
cosineplotdrugbank the ggplot object

cosineplotdrugbank the ggplot object
}
\description{
Creates the plot for all cosine coefficients against drugbank

Creates the plot for all cosine coefficients against drugbank
}
\examples{
\dontrun{
cosineepilepsyplot <- createCosinePlotAll(cosineepso, cosineesso)
}
\dontrun{
cosineepilepsyplot <- createCosinePlotAll(cosineepso, cosineesso)
}
}
