% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstDrugBank.R
\name{createCosinePlotDrugBank}
\alias{createCosinePlotDrugBank}
\title{Creates the plot for all cosine coefficients against drugbank}
\usage{
createCosinePlotDrugBank(dcosinedrugbank)
}
\arguments{
\item{dcosinedrugbank}{the data frame containing the columns with the cosine coefficients}
}
\value{
dcosineplotdrugbank the ggplot object
}
\description{
Creates the plot for all cosine coefficients against drugbank
}
\examples{
\dontrun{
createCosinePlotDrugBank(createCosineFrameDrugBank())
}
}
