% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstMeSH.R
\name{createCosinePlotMeSH}
\alias{createCosinePlotMeSH}
\title{Creates the plot for all cosine coefficients against mesh}
\usage{
createCosinePlotMeSH(cosinemesh)
}
\arguments{
\item{cosinemesh}{the data frame containing the columns with the cosine coefficients}
}
\value{
cosineplot the ggplot object
}
\description{
Creates the plot for all cosine coefficients against mesh
}
\examples{
\dontrun{
cosinemeshplot <- createCosinePlotMeSH(cosinemesh)
}
}
