% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstMeSH.R
\name{createDiceFrame}
\alias{createDiceFrame}
\title{Creates data frame for plotting the dice coefficients against MeSH}
\usage{
createDiceFrame()
}
\value{
ddice the data frame for the dice coefficient that can be used by createCosinePlot
}
\description{
Creates data frame for plotting the dice coefficients against MeSH
}
\examples{
\dontrun{
createDiceFrame()
}
}
