% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstDrugBank.R
\name{createDiceFrameDrugBank}
\alias{createDiceFrameDrugBank}
\title{Creates data frame for plotting the dice coefficients against DrugBank}
\usage{
createDiceFrameDrugBank(mesh, drugbank, epso, esso, epi)
}
\arguments{
\item{mesh}{list with mesh terms sorted by frequency}

\item{drugbank}{list with drugbank terms sorted by frequency}

\item{epso}{list with epso terms sorted by frequency}

\item{esso}{list with esso terms sorted by frequency}

\item{epi}{list with epi terms sorted by frequency}
}
\value{
dicedrugbank the data frame for the dice coefficient against DrugBank that can be used by createDicePlotDrugBank
}
\description{
Creates data frame for plotting the dice coefficients against DrugBank
}
\examples{
\dontrun{
dicedrugbank <- createDiceFrameDrugBank(mesh, drugbank, epso, esso, epi)
}
}
