% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsEpilepsyOntologies.R
\name{createDiceFrameEpSO}
\alias{createDiceFrameEpSO}
\title{Creates data frame for plotting the dice coefficients against EpSO}
\usage{
createDiceFrameEpSO(epso, esso, epi)
}
\arguments{
\item{epso}{list with epso terms sorted by frequency}

\item{esso}{list with esso terms sorted by frequency}

\item{epi}{list with epi terms sorted by frequency}
}
\value{
diceepso the data frame for the dice coefficient against EpSO
}
\description{
Creates data frame for plotting the dice coefficients against EpSO
}
\examples{
\dontrun{
diceepso <- createDiceFrameEpSO(epso, esso, epi)
}
}
