% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsEpilepsyOntologies.R
\name{createDiceFrameESSO}
\alias{createDiceFrameESSO}
\title{Creates data frame for plotting the dice coefficients against ESSO}
\usage{
createDiceFrameESSO(epso, esso, epi)
}
\arguments{
\item{epso}{list with epso terms sorted by frequency}

\item{esso}{list with esso terms sorted by frequency}

\item{epi}{list with epi terms sorted by frequency}
}
\value{
diceesso the data frame for the dice coefficient against ESSO
}
\description{
Creates data frame for plotting the dice coefficients against ESSO
}
\examples{
\dontrun{
diceesso <- createDiceFrameESSO(epso, esso, epi)
}
}
