% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{createDiceFrameMeSH}
\alias{createDiceFrameMeSH}
\title{Creates data frame for plotting the dice coefficients against MeSH}
\usage{
createDiceFrameMeSH(mesh, drugbank, epso, esso, epi)
}
\arguments{
\item{mesh}{list with mesh terms sorted by frequency}

\item{drugbank}{list with drugbank terms sorted by frequency}

\item{epso}{list with epso terms sorted by frequency}

\item{esso}{list with esso terms sorted by frequency}

\item{epi}{list with epi terms sorted by frequency}
}
\value{
ddicemesh the data frame for the dice coefficient against MeSH that can be used by createDicePlotMeSH
}
\description{
Creates data frame for plotting the dice coefficients against MeSH
}
\examples{
\dontrun{
ddicemesh <- createDiceFrameMeSH(mesh, drugbank, epso, esso, epi)
}
}
