% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstMeSH.R
\name{createDicePlot}
\alias{createDicePlot}
\title{Creates the plot for all dice coefficients against mesh}
\usage{
createDicePlot(ddice)
}
\arguments{
\item{ddice}{the data frame containing the columns with the dice coefficients}
}
\value{
diceplot the ggplot object
}
\description{
Creates the plot for all dice coefficients against mesh
}
\examples{
\dontrun{
createDicePlot(createDiceFrame)
}
}
