% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstAllEpilepsy.R
\name{createDicePlotAll}
\alias{createDicePlotAll}
\title{Creates the plot for all dice coefficients amongst the three epilepsy ontologies}
\usage{
createDicePlotAll(ddiceepso, ddiceesso)
}
\arguments{
\item{ddiceepso}{the data frame containing the columns with the dice coefficients of epi and esso against epso}

\item{ddiceesso}{the data frame containing the columns with the dice coefficients of epi and epso against esso}
}
\value{
diceepilepsyplot the ggplot object
}
\description{
Creates the plot for all dice coefficients amongst the three epilepsy ontologies
}
\examples{
\dontrun{
diceepilepsyplot <- createDicePlotAll(ddiceepso, ddiceesso)
}
}
