% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstDrugBank.R
\name{createDicePlotDrugBank}
\alias{createDicePlotDrugBank}
\title{Creates the plot for all dice coefficients against drugbank}
\usage{
createDicePlotDrugBank(dicedrugbank)
}
\arguments{
\item{dicedrugbank}{the data frame containing the columns with the dice coefficients}
}
\value{
diceplotdrugbank the ggplot object
}
\description{
Creates the plot for all dice coefficients against drugbank
}
\examples{
\dontrun{
dicedrugbankplot <- createDicePlotDrugBank(dicedrugbank)
}
}
