% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstMeSH.R
\name{createDicePlotMeSH}
\alias{createDicePlotMeSH}
\title{Creates the plot for all dice coefficients against mesh}
\usage{
createDicePlotMeSH(ddicemesh)
}
\arguments{
\item{ddicemesh}{the data frame containing the columns with the dice coefficients}
}
\value{
diceplotmesh the ggplot object
}
\description{
Creates the plot for all dice coefficients against mesh
}
\examples{
\dontrun{
createDicePlotMeSH(createDiceFrameMeSH())
}
}
