% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstMeSH.R
\name{createJaccardFrame}
\alias{createJaccardFrame}
\title{Creates data frame for plotting the jaccard coefficients against MeSH}
\usage{
createJaccardFrame()
}
\value{
dcosine the data frame for the jaccard coefficient that can be used by createCosinePlot
}
\description{
Creates data frame for plotting the jaccard coefficients against MeSH
}
\examples{
\dontrun{
createJaccardFrame()
}
}
