% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstDrugBank.R
\name{createJaccardFrameDrugBank}
\alias{createJaccardFrameDrugBank}
\title{Creates data frame for plotting the jaccard coefficients against DrugBank}
\usage{
createJaccardFrameDrugBank(mesh, drugbank, epso, esso, epi)
}
\value{
djaccarddrugbank the data frame for the jaccard coefficient against DrugBank that can be used by createJaccardPlotDrugBank
}
\description{
Creates data frame for plotting the jaccard coefficients against DrugBank
}
\examples{
\dontrun{
createJaccardFrameDrugBank(mesh, drugbank, epso, esso, epi)
}
}
