% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsEpilepsyOntologies.R
\name{createJaccardFrameEpSO}
\alias{createJaccardFrameEpSO}
\title{Creates data frame for plotting the jaccard coefficients against EpSO}
\usage{
createJaccardFrameEpSO(epso, esso, epi)
}
\arguments{
\item{epso}{list with epso terms sorted by frequency}

\item{esso}{list with esso terms sorted by frequency}

\item{epi}{list with epi terms sorted by frequency}
}
\value{
djaccardepso the data frame for the jaccard coefficient against EpSO
}
\description{
Creates data frame for plotting the jaccard coefficients against EpSO
}
\examples{
\dontrun{
djaccardepso <- createJaccardFrameEpSO(epso, esso, epi)
}
}
