% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsEpilepsyOntologies.R
\name{createJaccardFrameESSO}
\alias{createJaccardFrameESSO}
\title{Creates data frame for plotting the jaccard coefficients against ESSO}
\usage{
createJaccardFrameESSO()
}
\value{
jaccardesso the data frame for the jaccard coefficient against ESSO
}
\description{
Creates data frame for plotting the jaccard coefficients against ESSO
}
\examples{
jaccardesso <- createJaccardFrameESSO()

}
