% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{createJaccardFrameMeSH}
\alias{createJaccardFrameMeSH}
\title{Creates data frame for plotting the jaccard coefficients against MeSH}
\usage{
createJaccardFrameMeSH(mesh, drugbank, epso, esso, epi)
}
\value{
djaccardmesh the data frame for the jaccard coefficient that can be used by createJaccardPlotMeSH
}
\description{
Creates data frame for plotting the jaccard coefficients against MeSH
}
\examples{
\dontrun{
createJaccardFrameMeSH()
}
}
