% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstMeSH.R
\name{createJaccardPlot}
\alias{createJaccardPlot}
\title{Creates the plot for all jaccard coefficients against mesh}
\usage{
createJaccardPlot(djaccard)
}
\arguments{
\item{djaccard}{the data frame containing the columns with the jaccard coefficients}
}
\value{
jaccardplot the ggplot object
}
\description{
Creates the plot for all jaccard coefficients against mesh
}
\examples{
\dontrun{
createJaccardPlot(createJaccardFrame)
}
}
