% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstAllEpilepsy.R
\name{createJaccardPlotAll}
\alias{createJaccardPlotAll}
\title{Creates the plot for all jaccard coefficients amongst the three epilepsy ontologies}
\usage{
createJaccardPlotAll(djaccardepso, djaccardesso)
}
\arguments{
\item{djaccardepso}{the data frame containing the columns with the jaccard coefficients of epi and esso against epso}

\item{djaccardesso}{the data frame containing the columns with the jaccard coefficients of epi and epso against esso}
}
\value{
jaccardepilepsyplot the ggplot object
}
\description{
Creates the plot for all jaccard coefficients amongst the three epilepsy ontologies
}
\examples{
\dontrun{
jaccardepilepsyplot <- createJaccardPlotAll(djaccardepso, djaccardesso)
}
}
