% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotDrugVectors.R
\name{createJaccardPlotDBMeSH}
\alias{createJaccardPlotDBMeSH}
\title{Creates the plot for all jaccard coefficients amongst the three epilepsy ontologies}
\usage{
createJaccardPlotDBMeSH(jmeshepso, jmeshesso, jmeshepi)
}
\arguments{
\item{jmeshepso}{list containing jaccard coefficients between mesh and epso for increasing k}

\item{jmeshesso}{list containing jaccard coefficients between mesh and esso for increasing k}

\item{jmeshepi}{list containing jaccard coefficients between mesh and epi for increasing k}
}
\value{
jaccardepilepsyplot the ggplot object
}
\description{
Creates the plot for all jaccard coefficients amongst the three epilepsy ontologies
}
\examples{
\dontrun{
jaccardepilepsyplot <- createJaccardPlotAll(jaccardepso, jaccardesso)
}
}
