% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstDrugBank.R
\name{createJaccardPlotDrugBank}
\alias{createJaccardPlotDrugBank}
\title{Creates the plot for all jaccard coefficients against drugbank}
\usage{
createJaccardPlotDrugBank(jaccarddrugbank)
}
\arguments{
\item{jaccarddrugbank}{the data frame containing the columns with the jaccard coefficients}
}
\value{
jaccardplotdrugbank the ggplot object
}
\description{
Creates the plot for all jaccard coefficients against drugbank
}
\examples{
\dontrun{
jaccarddrugbankplot <- createJaccardPlotDrugBank(jaccarddrugbank)
}
}
