% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstMeSH.R
\name{createJaccardPlotMeSH}
\alias{createJaccardPlotMeSH}
\title{Creates the plot for all jaccard coefficients against mesh}
\usage{
createJaccardPlotMeSH(jaccardmesh)
}
\arguments{
\item{jaccardmesh}{the data frame containing the columns with the jaccard coefficients}
}
\value{
jaccardmeshplot the ggplot object
}
\description{
Creates the plot for all jaccard coefficients against mesh
}
\examples{
\dontrun{
jaccardmeshplot <- createJaccardPlotMeSH(jaccardmesh)
}
}
