% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotDrugVectors.R
\name{createJaccardPlotMeSHFive}
\alias{createJaccardPlotMeSHFive}
\title{Creates the plot for all jaccard coefficients amongst the three epilepsy ontologies}
\usage{
createJaccardPlotMeSHFive(jmeshepso, jmeshesso, jmeshepi, jmeshepilepsyand,
  jmeshepilepsyor)
}
\arguments{
\item{jmeshepso}{list of jaccard coefficients between mesh and epso for increasing k}

\item{jmeshesso}{list of jaccard coefficients between mesh and esso for increasing k}

\item{jmeshepi}{list of jaccard coefficients between mesh and epi for increasing k}

\item{jmeshepilepsyand}{list of jaccard coefficients between mesh and the intersection of epso, esso, and epi for increasing k}

\item{jmeshepilepsyor}{list of jaccard coefficients between mesh and the union of epso, esso, and epi for increasing k}
}
\value{
jaccardepilepsyplot the ggplot object
}
\description{
Creates the plot for all jaccard coefficients amongst the three epilepsy ontologies
}
\examples{
\dontrun{
jaccardepilepsyplot <- createJaccardPlotAll(jaccardepso, jaccardesso)
}
}
