% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResults.R
\name{createListFromVennVector}
\alias{createListFromVennVector}
\title{Creates a list out of a vector}
\usage{
createListFromVennVector(vennvector)
}
\arguments{
\item{vennvector}{vector containing}
}
\value{
res list with elements from the vector
}
\description{
Creates a list out of a vector
}
\examples{
\dontrun{
createListFromVennVector(vennvector)
}
}
