% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSimilarities.R
\name{dice}
\alias{dice}
\title{Calculate dice similarity metric}
\usage{
dice(ainterb, lengtha, lengthb)
}
\arguments{
\item{ainterb}{integer value with number of intersecting elements between set a and b}

\item{lengtha}{integer value with the number of items in set a}

\item{lengthb}{integer value with the number of items in set b}
}
\value{
dice double vlaue with the dice similarity coefficient
}
\description{
Calculate dice similarity metric
}
\examples{
dice(1, 3, 4)
}
