% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridArrangeCall.R
\name{doFullPlot}
\alias{doFullPlot}
\title{Does the full plot on one page}
\usage{
doFullPlot(
  cosinemeshplot,
  cosinedrugbankplot,
  cosineepilepsyplot,
  dicemeshplot,
  dicedrugbankplot,
  diceepilepsyplot,
  jaccardmeshplot,
  jaccarddrugbankplot,
  jaccardepilepsyplot
)
}
\arguments{
\item{cosinemeshplot}{plot with cosine coefficients against MeSH}

\item{cosinedrugbankplot}{plot with cosine coefficients against DrugBank}

\item{cosineepilepsyplot}{plot with cosine coefficients of Epilepsy Ontologies}

\item{dicemeshplot}{plot with dice coefficients against MeSH}

\item{dicedrugbankplot}{plot with dice coefficients against DrugBank}

\item{diceepilepsyplot}{plot with dice coefficients of Epilepsy Ontologies}

\item{jaccardmeshplot}{plot with jaccard coefficients against MeSH}

\item{jaccarddrugbankplot}{plot with jaccard coefficients against DrugBank}

\item{jaccardepilepsyplot}{plot with jaccard coefficients of Epilepsy Ontologies}
}
\value{
full
}
\description{
Does the full plot on one page
}
\examples{
\dontrun{
full <- doFullPlot (cosinemeshplot, 
                    cosinedrugbankplot, 
                    cosineepilepsyplot, 
                    dicemeshplot, 
                    dicedrugbankplot, 
                    diceepilepsyplot, 
                    jaccardmeshplot, 
                    jaccarddrugbankplot, 
                    jaccardepilepsyplot)
}
}
