% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVennDiagramms.R
\name{drawVenn4}
\alias{drawVenn4}
\title{Create Venn Diagramm for the set properties of four dictionaries from the
source ontologies}
\usage{
drawVenn4()
}
\value{
plot object
}
\description{
Create Venn Diagramm for the set properties of four dictionaries from the
source ontologies
}
\examples{
ggplot2::ggsave("venn4.png", plot = drawVenn4(), width=240, height=160, 
  units = "mm", dpi = 300)
}
