% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVennDiagramms.R
\name{drawVenn4Doc}
\alias{drawVenn4Doc}
\title{Create quintuple Venn Diagramm for shared documents with co-occurrences of
drug names between EpSO, ESSO, EPILONT and EPISEM}
\usage{
drawVenn4Doc()
}
\value{
plot object
}
\description{
Create quintuple Venn Diagramm for shared documents with co-occurrences of
drug names between EpSO, ESSO, EPILONT and EPISEM
}
\examples{
\dontrun{
ggplot2::ggsave("venn4doc.png", plot = drawVenn4Doc(), width=240, height=160, 
  units = "mm", dpi = 300)
}
}
