% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVennDiagramms.R
\name{drawVenn4Syn}
\alias{drawVenn4Syn}
\title{Create quad Venn Diagramm for shared synonyms between EpSO, ESSO, EPILONT
and EPISEM}
\usage{
drawVenn4Syn()
}
\value{
plot object
}
\description{
Create quad Venn Diagramm for shared synonyms between EpSO, ESSO, EPILONT
and EPISEM
}
\examples{
\dontrun{
ggplot2::ggsave("venn4syn.png", plot = drawVenn4Syn(), width=240, 
  height=160, units = "mm", dpi = 300)
}
}
