% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVennDiagramms.R
\name{drawVenn5}
\alias{drawVenn5}
\title{Create Venn Diagramm for the set properties of all five dictionaries from the
source ontologies}
\usage{
drawVenn5()
}
\value{
plot object
}
\description{
Create Venn Diagramm for the set properties of all five dictionaries from the
source ontologies
}
\examples{
\dontrun{
ggplot2::ggsave("venn5.png", plot = drawVenn5(), width=240, height=160, 
  units = "mm", dpi = 300)
}
}
