% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVennDiagramms.R
\name{drawVenn5Doc}
\alias{drawVenn5Doc}
\title{Create quintuple Venn Diagramm for shared documents between EpSO, ESSO, 
EPILONT, EPISEM and FENICS}
\usage{
drawVenn5Doc()
}
\value{
plot object
}
\description{
Create quintuple Venn Diagramm for shared documents between EpSO, ESSO, 
EPILONT, EPISEM and FENICS
}
\examples{
\dontrun{
ggplot2::ggsave("venn5doc.png", plot = drawVenn5Doc(), width=240, height=160, 
  units = "mm", dpi = 300)
}
}
