% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVennDiagramms.R
\name{drawVenn5DrugDoc}
\alias{drawVenn5DrugDoc}
\title{Create quintuple Venn Diagramm for shared documents with co-occurrences of
drug names between EpSO, ESSO, EPILONT, EPISEM and FENICS}
\usage{
drawVenn5DrugDoc()
}
\value{
plot object
}
\description{
Create quintuple Venn Diagramm for shared documents with co-occurrences of
drug names between EpSO, ESSO, EPILONT, EPISEM and FENICS
}
\examples{
\dontrun{
ggplot2::ggsave("venn5drugdoc.png", plot = drawVenn5DrugDoc(), width=240, 
  height=160, units = "mm", dpi = 300)
}
}
