% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVennDiagramms.R
\name{drawVenn5Syn}
\alias{drawVenn5Syn}
\title{Create quintuple Venn Diagramm for shared synonyms between EpSO, ESSO, 
EPILONT, EPISEM and FENICS}
\usage{
drawVenn5Syn()
}
\value{
plot object
}
\description{
Create quintuple Venn Diagramm for shared synonyms between EpSO, ESSO, 
EPILONT, EPISEM and FENICS
}
\examples{
\dontrun{
ggplot2::ggsave("venn5syn.png", plot = drawVenn5Syn(), width=240, 
  height=160, units = "mm", dpi = 300)
}
}
