% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVennDiagramms.R
\name{drawVennGrid}
\alias{drawVennGrid}
\title{Create plot_grid from multiple plots}
\usage{
drawVennGrid()
}
\value{
plot object
}
\description{
Create plot_grid from multiple plots
}
\examples{
\dontrun{
  cowplot::plot_grid(drawVenn4 (), drawVenn4Syn(), drawVenn5Doc (), 
    drawVenn5DrugDoc ())
  ggplot2::ggsave("vennAB.png", plot = cowplot::plot_grid(drawVenn4 (), 
    drawVenn4Syn(), labels = c('A', 'B'), ncol = 1), width=240, height=320,  
    units = "mm", dpi = 300)
  ggplot2::ggsave("vennAB.png", plot = cowplot::plot_grid(drawVenn4 (), 
    drawVenn4Syn(), labels = c('Concepts:', 'Synonyms:'), ncol = 1), width=240, 
    height=320,  units = "mm", dpi = 300)
  ggplot2::ggsave("vennCD.png", plot = cowplot::plot_grid(drawVenn5Doc (), 
    drawVenn5DrugDoc(), labels = c('Documents with B-Terms:', 
    'Documents with B- and C-Terms:'), ncol = 1), width=240, height=320,  
    units = "mm", dpi = 300)
  ggplot2::ggsave("vennCD.png", plot = cowplot::plot_grid(drawVenn5Doc (), 
    drawVenn5DrugDoc(), labels = c('Documents with B-Terms:', 
   'Documents with B- and C-Terms:'), ncol = 1), width=240, height=320,  units = "mm", 
   dpi = 300)
  ggplot2::ggsave("vennCD.png", plot = cowplot::plot_grid(drawVenn4Doc (), 
    drawVenn4DrugDoc(), labels = c('Documents with B-Terms:', 
    'Documents with B- and C-Terms:'), ncol = 1), width=240, height=320,  
    units = "mm", dpi = 300)
  ggplot2::ggsave("vennCD.png", plot = cowplot::plot_grid(drawVenn4Doc (), 
    drawVenn4DrugDoc(), labels = c('Documents\nwith B-Terms:             ', 
    'Documents\nwith B- and C-Terms:'), ncol = 1), width=240, height=320,  
    units = "mm", dpi = 300)
  ggplot2::ggsave("vennAB.png", plot = cowplot::plot_grid(drawVenn4 (), 
    drawVenn4Syn(), labels = c('i) Concepts:', 'ii) Synonyms:'), ncol = 1), 
    width=240, height=320,  units = "mm", dpi = 300)
  ggplot2::ggsave("vennCD.png", plot = cowplot::plot_grid(NULL, 
    drawVenn4Doc (), drawVenn4DrugDoc(), 
    labels = c('iii) Documents with B-Terms:', 
    'iv) Documents with B- and C-Terms:'), ncol = 1, 
    label_x = c(-0.105, -0.14), label_fontfamily = "Arial Nova Light", 
    label_fontface = "bold"), width=240, height=320,  units = "mm", dpi = 300)
}
}
