% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{drawVennPlot}
\alias{drawVennPlot}
\title{Draw venn diagram with example values}
\usage{
drawVennPlot(dneuromaxk)
}
\arguments{
\item{dneuromaxk}{object created with TopKLists::calculate.maxK()}
}
\value{
venn diagram
}
\description{
Draw venn diagram with example values
}
\examples{
utils::data(rawDrugBankCoOcEpSO, package="epos")
atchashda <-
  readAtcMapIntoHashMapDrugNamesAtcCodes(
    system.file("extdata", "db-atc.map", package = "epos"), "\\t")
tepso <- genDictListFromRawFreq(rawDrugBankCoOcEpSO)
neuroepso <- filterNeuroDrugs(tepso, atchashda)
utils::data(rawDrugBankCoOcESSO, package="epos")
tesso <- genDictListFromRawFreq(rawDrugBankCoOcESSO)
neuroesso <- filterNeuroDrugs(tesso, atchashda)
utils::data(rawDrugBankCoOcEPILONT, package="epos")
tepi <- genDictListFromRawFreq(rawDrugBankCoOcEPILONT)
neuroepi <- filterNeuroDrugs(tepi, atchashda)
dneuro <-
  data.frame(EpSO = neuroepso[1:210],
             ESSO = neuroesso[1:210],
             EPILONT = neuroepi[1:210])
dneuromaxk <- TopKLists::calculate.maxK(dneuro, 3, 5, 10)
v <- drawVennPlot(dneuromaxk)
}
