% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{getPerms}
\alias{getPerms}
\title{Aggregat a list to matrix}
\usage{
getPerms(x)
}
\arguments{
\item{x}{this is a list}
}
\value{
res this is a matrix
}
\description{
Aggregat a list to matrix
}
\examples{
\dontrun{
getPerms(c("a","b"))
}
}
