% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnrichment.R
\name{getRefAll}
\alias{getRefAll}
\title{Retrieve the list of drugs from the union of all reference lists}
\usage{
getRefAll()
}
\value{
list of drugs from all reference lists
}
\description{
Retrieve the list of drugs from the union of all reference lists
}
\examples{
d <- getRefAll()
}
