% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDictionaryStatsFromMongo.R
\name{getTermMatrix}
\alias{getTermMatrix}
\title{Receives a sorted hashmap with found entities from a dictionary}
\usage{
getTermMatrix(dictionary, collection)
}
\arguments{
\item{dictionary}{Character vector that is the name of a dicitonary having pre-calculated stats. This can be
MeSH, DrugBank, Agrovoc, EpSO, ESSO, or EPILONT}
}
\value{
a sorted hashmap containing all found entities from the respective dictionaries with frequencies
}
\description{
Receives a sorted hashmap with found entities from a dictionary
}
\examples{
\dontrun{
mesh <- getTermMatrix("MeSH")
}

}
