% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDictFreq.R
\name{loadDictionaryFrequencyDrugBankCoOcEpSO}
\alias{loadDictionaryFrequencyDrugBankCoOcEpSO}
\title{Load the RDS object inst/extdata/DrugBankCoOcEpSO.rds into a variable that is returned by this function. 
The returned list contains all terms from the dictionary DrugBank that co-occur in the same
documents with terms from the EpSO dictionary in the corpus of LIVIVO. 
The moment of execution for generating the data 
stored in the RDS-file on the corpus of LIVIVO is June 26th 2018.}
\usage{
loadDictionaryFrequencyDrugBankCoOcEpSO()
}
\value{
tepso list with drug terms sorted by their frequency
}
\description{
Load the RDS object inst/extdata/DrugBankCoOcEpSO.rds into a variable that is returned by this function. 
The returned list contains all terms from the dictionary DrugBank that co-occur in the same
documents with terms from the EpSO dictionary in the corpus of LIVIVO. 
The moment of execution for generating the data 
stored in the RDS-file on the corpus of LIVIVO is June 26th 2018.
}
\examples{
tepso <- loadDictionaryFrequencyDrugBankCoOcEpSO ()

}
