% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDictFreq.R
\name{loadDictionaryFrequencyESSO}
\alias{loadDictionaryFrequencyESSO}
\title{Load the RDS object inst/resources/ESSO.rds into a variable that is returned by this function. 
The returned list contains all terms from the dictionary ESSO with its frequency of
occurrences in the LIVIVO corpus. The moment of execution for generating the data 
stored in the RDS-file on the corpus of LIVIVO is June 26th 2018.}
\usage{
loadDictionaryFrequencyESSO()
}
\value{
mesh list with ESSO terms sorted by their frequency
}
\description{
Load the RDS object inst/resources/ESSO.rds into a variable that is returned by this function. 
The returned list contains all terms from the dictionary ESSO with its frequency of
occurrences in the LIVIVO corpus. The moment of execution for generating the data 
stored in the RDS-file on the corpus of LIVIVO is June 26th 2018.
}
\examples{
esso <- loadDictionaryFrequencyESSO ()
}
