% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOntStatsIntoEnv.R
\name{loadDrugBank}
\alias{loadDrugBank}
\title{Loads drugbank object from harddrive into the R environment}
\usage{
loadDrugBank()
}
\value{
adrugbank list of drugbank terms sorted by frequency
}
\description{
Loads drugbank object from harddrive into the R environment
}
\examples{
\dontrun{
loadDrugBank()
}
}
