% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOntStatsIntoEnv.R
\name{loadEPILONT}
\alias{loadEPILONT}
\title{Loads epilont object from harddrive into the R environment}
\usage{
loadEPILONT()
}
\value{
aepi list of mesh terms sorted by frequency
}
\description{
Loads epilont object from harddrive into the R environment
}
\examples{
\dontrun{
loadEPILONT()
}
}
