% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOntStatsIntoEnv.R
\name{loadEpSO}
\alias{loadEpSO}
\title{Loads epso object from harddrive into the R environment}
\usage{
loadEpSO()
}
\value{
aepso list of mesh terms sorted by frequency
}
\description{
Loads epso object from harddrive into the R environment
}
\examples{
\dontrun{
loadEpSO()
}
}
