% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOntStatsIntoEnv.R
\name{loadESSO}
\alias{loadESSO}
\title{Loads esso object from harddrive into the R environment}
\usage{
loadESSO()
}
\value{
aesso list of mesh terms sorted by frequency
}
\description{
Loads esso object from harddrive into the R environment
}
\examples{
\dontrun{
loadESSO()
}
}
