% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOntStatsIntoEnv.R
\name{loadMeSH}
\alias{loadMeSH}
\title{Loads mesh object from harddrive into the R environment}
\usage{
loadMeSH()
}
\value{
amesh list of mesh terms sorted by frequency
}
\description{
Loads mesh object from harddrive into the R environment
}
\examples{
\dontrun{
loadMeSH()
}
}
