% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOntStatsIntoEnv.R
\name{loadSavedSortedMapsIntoEnv}
\alias{loadSavedSortedMapsIntoEnv}
\title{Loads objects from harddrive into the R environment}
\usage{
loadSavedSortedMapsIntoEnv()
}
\value{
no direct return but loading into environment
}
\description{
Loads objects from harddrive into the R environment
}
\examples{
\dontrun{
loadSavedSortedMapsIntoEnv()
}
}
