% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main function to call everything and produce the results}
\usage{
main()
}
\value{
result table
}
\description{
Main function to call everything and produce the results
}
\examples{
main()
}
