% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnrichment.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plotting functions for enrichment lists}
\usage{
plotEnrichment(enepso, enesso, enepi, enepisem, enfenics)
}
\arguments{
\item{enepso}{list with enrichment for EpSO}

\item{enesso}{list with enrichment for ESSO}

\item{enepi}{list with enrichment for EPILONT}

\item{enepisem}{list with enrichment for EPISEM}

\item{enfenics}{list with enrichment for FENICS}
}
\value{
the plot object
}
\description{
Plotting functions for enrichment lists
}
\examples{
plotEnrichment (c(0,1,2,2,3),
                c(0,0,1,1,2),
                c(0,1,1,2,2),
                c(0,1,2,3,3),
                c(1,2,3,3,3))
}
