% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processATC.R
\name{readAtcMapIntoHashMapAtcCodesAtcNames}
\alias{readAtcMapIntoHashMapAtcCodesAtcNames}
\title{Processes the input file db-atc.map to form a HashMap containing the drug names with ATC codes}
\usage{
readAtcMapIntoHashMapAtcCodesAtcNames(filename, seperator)
}
\arguments{
\item{filename}{character vector with the file name of the file db-atc.map}

\item{seperator}{character vector with the seperator used within the map-file}
}
\value{
atchashaa hashmap with atc codes as keys and atc names as values
}
\description{
Processes the input file db-atc.map to form a HashMap containing the drug names with ATC codes
}
\examples{
atchashaa <-
  readAtcMapIntoHashMapAtcCodesAtcNames(
    system.file("extdata", "db-atc.map", package = "epos"), "\\t")
}
