% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processATC.R
\name{readAtcMapIntoHashMapDrugNamesAtcCodes}
\alias{readAtcMapIntoHashMapDrugNamesAtcCodes}
\title{Processes the input file db-atc.map to form a HashMap containing the drug names with ATC codes}
\usage{
readAtcMapIntoHashMapDrugNamesAtcCodes(filename, seperator)
}
\arguments{
\item{filename}{character vector with the file name of the file db-atc.map}

\item{seperator}{character vector with the seperator used within the map-file}
}
\value{
atchashda hashmap with drug names as keys and atc codes as values
}
\description{
Processes the input file db-atc.map to form a HashMap containing the drug names with ATC codes
}
\examples{
atchashda <- readAtcMapIntoHashMapDrugNamesAtcCodes(
  system.file("extdata", "db-atc.map", package = "epos"), "\\t")
}
