% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processATC.R
\name{readThirdLevelATC}
\alias{readThirdLevelATC}
\title{Read the third level ATC classes from the file drugbankatc-thirdlevel.map}
\usage{
readThirdLevelATC(filename, seperator)
}
\arguments{
\item{filename}{the file name that is supposed to be drugbankatc-thirdlevel.map}

\item{seperator}{the csv file delimiter}
}
\value{
atchashthird a hashmap with third level ATC classes as keys and their names as values
}
\description{
Read the third level ATC classes from the file drugbankatc-thirdlevel.map
}
\examples{
\dontrun{
atchashthird <- readThirdLevelATC("inst/resources/drugbankatc-thirdlevel.map", "\\t")
}
}
