% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processATC.R
\name{sortHashMapByValue}
\alias{sortHashMapByValue}
\title{Sort a hashmap by its values}
\usage{
sortHashMapByValue(ahashmap)
}
\arguments{
\item{ahashmap}{a hashmap that will be sorted by its values}
}
\value{
sortedhm the sorted hashmap
}
\description{
Sort a hashmap by its values
}
\examples{
\dontrun{
shm <- sortHashMapByValue(ahm)
}
}
