% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAggregatedEpilepsy.R
\name{aggregatedPlotEpilepsie}
\alias{aggregatedPlotEpilepsie}
\title{Title}
\usage{
aggregatedPlotEpilepsie(cosineepso, cosineesso, diceepso, diceesso, jaccardesso,
  jaccardepso)
}
\arguments{
\item{cosineepso}{}

\item{cosineesso}{}

\item{diceepso}{}

\item{diceesso}{}

\item{jaccardesso}{}

\item{jaccardepso}{}
}
\value{
aggeps
}
\description{
Title
}
\examples{
\dontrun{
aggeps <- aggregatedPlotEpilepsie(cosineepso, cosineesso, diceepso, diceesso, jaccardesso, jaccardepso)
}
}
