% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSimilarities.R
\name{calcDice}
\alias{calcDice}
\title{Calculate the dice for two sets a and b}
\usage{
calcDice(a, b)
}
\arguments{
\item{a}{list with elements that should be of same type as in list b}

\item{b}{list with elements}
}
\value{
di list with length of longer set a or b containing the dice similarity coefficient at each position
}
\description{
Calculate the dice for two sets a and b
}
\examples{
\dontrun{
calcDice(c(1,2,3), c(2,3,4))
}
}
