% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processATC.R
\name{countATC}
\alias{countATC}
\title{Count the ATC second level classes for all drug names in the druglist}
\usage{
countATC(druglist, atchashda, atchashaa, atchashsec)
}
\arguments{
\item{druglist}{a character vector containing a list of drug names}

\item{atchashda}{a hashmap with drug names as keys and atc classes as values}

\item{atchashaa}{a hashmap with atc classes as keys and atc names as values}

\item{atchashsec}{a hashmap containing the second level atc classes as keys and their names as values}
}
\value{
atccounter
}
\description{
Count the ATC second level classes for all drug names in the druglist
}
\examples{
\dontrun{
atccounter <- countATC(druglist, atchashda, atchashaa, atchashsec)
}
}
