% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResults.R
\name{createBaseTable}
\alias{createBaseTable}
\title{Create the final resulting data frame}
\usage{
createBaseTable(neurospace, atchashda, atchashsec, dneuromaxk)
}
\arguments{
\item{neurospace}{list of drug names that were aggregated using TopKLists::calculate.max topkspace}

\item{atchashda}{hashmap retrieved from readAtcMapIntoHashMapDrugNamesAtcCodes}

\item{atchashsec}{hashmap retrieved from readSecondLevelATC}

\item{dneuromaxk}{data frame containing columns for each intersection, ATC class, and reference list}
}
\value{
data frame with all resuults
}
\description{
Create the final resulting data frame
}
\examples{
\dontrun{
createBaseTable(neurospace, atchashda, atchashsec, dneuromaxk)
}
}
