% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsEpilepsyOntologies.R
\name{createCosineFrameESSO}
\alias{createCosineFrameESSO}
\title{Creates data frame for plotting the cosine coefficients against ESSO}
\usage{
createCosineFrameESSO()
}
\value{
cosineesso the data frame for the cosine coefficient against ESSO
}
\description{
Creates data frame for plotting the cosine coefficients against ESSO
}
\examples{
cosineesso <- createCosineFrameESSO()

}
