% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{createCosineFrameMeSH}
\alias{createCosineFrameMeSH}
\title{Creates data frame for plotting the cosine coefficients against MeSH}
\usage{
createCosineFrameMeSH(mesh, drugbank, epso, esso, epi)
}
\arguments{
\item{mesh}{list with mesh terms sorted by frequency}

\item{drugbank}{list with drugbank terms sorted by frequency}

\item{epso}{list with epso terms sorted by frequency}

\item{esso}{list with esso terms sorted by frequency}

\item{epi}{list with epi terms sorted by frequency}
}
\value{
dcosinemesh the data frame for the cosine coefficient against MeSH that can be used by createCosinePlotMeSH
}
\description{
Creates data frame for plotting the cosine coefficients against MeSH
}
\examples{
\dontrun{
dcosinemesh <- createCosineFrameMeSH(mesh, drugbank, epso, esso, epi)
}
}
