% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstAllEpilepsy.R
\name{createCosinePlotAll}
\alias{createCosinePlotAll}
\title{Creates the plot for all cosine coefficients against drugbank}
\usage{
createCosinePlotAll(dcosineepso, dcosineesso)
}
\arguments{
\item{dcosineepso}{the data frame containing the columns with the cosine coefficients of epi and esso against epso}

\item{dcosineesso}{the data frame containing the columns with the cosine coefficients of epi and epso against esso}
}
\value{
cosineplotdrugbank the ggplot object
}
\description{
Creates the plot for all cosine coefficients against drugbank
}
\examples{
\dontrun{
cosineepilepsyplot <- createCosinePlotAll(dcosineepso, dcosineesso)
}
}
