% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstDrugBank.R
\name{createDiceFrameDrugBank}
\alias{createDiceFrameDrugBank}
\title{Creates data frame for plotting the dice coefficients against DrugBank}
\usage{
createDiceFrameDrugBank(mesh, drugbank, epso, esso, epi)
}
\value{
ddicedrugbank the data frame for the dice coefficient against DrugBank that can be used by createDicePlotDrugBank
}
\description{
Creates data frame for plotting the dice coefficients against DrugBank
}
\examples{
\dontrun{
createDiceFrameDrugBank(mesh, drugbank, epso, esso, epi)
}
}
