% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsEpilepsyOntologies.R
\name{createDiceFrameESSO}
\alias{createDiceFrameESSO}
\title{Creates data frame for plotting the dice coefficients against ESSO}
\usage{
createDiceFrameESSO()
}
\value{
diceesso the data frame for the dice coefficient against ESSO
}
\description{
Creates data frame for plotting the dice coefficients against ESSO
}
\examples{
diceesso <- createDiceFrameESSO()

}
