% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsAgainstMeSH.R
\name{createDiceFrameMeSH}
\alias{createDiceFrameMeSH}
\title{Creates data frame for plotting the dice coefficients against MeSH}
\usage{
createDiceFrameMeSH(mesh, drugbank, epso, esso, epi)
}
\value{
ddicemesh the data frame for the dice coefficient against MeSH that can be used by createDicePlotMeSH
}
\description{
Creates data frame for plotting the dice coefficients against MeSH
}
\examples{
\dontrun{
createDiceFrameMeSH()
}
}
