% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstAllEpilepsy.R,
%   R/doThePlotAgainstBaseline_DB-MeSH.R
\name{createDicePlotAll}
\alias{createDicePlotAll}
\alias{createDicePlotAll}
\title{Creates the plot for all dice coefficients amongst the three epilepsy ontologies}
\usage{
createDicePlotAll(diceepso, diceesso)

createDicePlotAll(diceepso, diceesso)
}
\arguments{
\item{diceepso}{the data frame containing the columns with the dice coefficients of epi and esso against epso}

\item{diceesso}{the data frame containing the columns with the dice coefficients of epi and epso against esso}

\item{diceepso}{the data frame containing the columns with the dice coefficients of epi and esso against epso}

\item{diceesso}{the data frame containing the columns with the dice coefficients of epi and epso against esso}
}
\value{
diceepilepsyplot the ggplot object

diceepilepsyplot the ggplot object
}
\description{
Creates the plot for all dice coefficients amongst the three epilepsy ontologies

Creates the plot for all dice coefficients amongst the three epilepsy ontologies
}
\examples{
\dontrun{
diceepilepsyplot <- createDicePlotAll(diceepso, diceesso)
}
\dontrun{
diceepilepsyplot <- createDicePlotAll(diceepso, diceesso)
}
}
