% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsEpilepsyOntologies.R
\name{createJaccardFrameESSO}
\alias{createJaccardFrameESSO}
\title{Creates data frame for plotting the jaccard coefficients against ESSO}
\usage{
createJaccardFrameESSO(epso, esso, epi)
}
\arguments{
\item{epso}{list with epso terms sorted by frequency}

\item{esso}{list with esso terms sorted by frequency}

\item{epi}{list with epi terms sorted by frequency}
}
\value{
djaccardesso the data frame for the jaccard coefficient against ESSO
}
\description{
Creates data frame for plotting the jaccard coefficients against ESSO
}
\examples{
\dontrun{
djaccardesso <- createJaccardFrameESSO(epso, esso, epi)
}
}
