% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstBaseline_DB-MeSH.R
\name{createJaccardPlotDBMeSH}
\alias{createJaccardPlotDBMeSH}
\title{Creates the plot for all jaccard coefficients amongst the three epilepsy ontologies}
\usage{
createJaccardPlotDBMeSH(jmeshepso, jmeshesso, jmeshepi)
}
\arguments{
\item{jaccardepso}{the data frame containing the columns with the jaccard coefficients of epi and esso against epso}

\item{jaccardesso}{the data frame containing the columns with the jaccard coefficients of epi and epso against esso}
}
\value{
jaccardepilepsyplot the ggplot object
}
\description{
Creates the plot for all jaccard coefficients amongst the three epilepsy ontologies
}
\examples{
\dontrun{
jaccardepilepsyplot <- createJaccardPlotAll(jaccardepso, jaccardesso)
}
}
