% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotAgainstMeSH.R
\name{createJaccardPlotMeSH}
\alias{createJaccardPlotMeSH}
\title{Creates the plot for all jaccard coefficients against mesh}
\usage{
createJaccardPlotMeSH(djaccardmesh)
}
\arguments{
\item{djaccardmesh}{the data frame containing the columns with the jaccard coefficients}
}
\value{
jaccardplotmesh the ggplot object
}
\description{
Creates the plot for all jaccard coefficients against mesh
}
\examples{
\dontrun{
createJaccardPlotMeSH(createJaccardFrameMeSH())
}
}
