% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doThePlotDrugVectors.R
\name{createTanimotoBaseline}
\alias{createTanimotoBaseline}
\title{Creates the plot for all jaccard coefficients amongst the three epilepsy ontologies}
\usage{
createTanimotoBaseline(neuroepso, neuroesso, neuroepi, dneuromaxk)
}
\arguments{
\item{neuroepso}{list of neuro drug names co-occurring with epso}

\item{neuroesso}{list of neuro drug names co-occurring with esso}

\item{neuroepi}{list of neuro drug names co-occurring with epi}

\item{dneuromaxk}{object returned from TopKLists::calculate.maxkS}
}
\value{
jaccardepilepsyplot the ggplot object
}
\description{
Creates the plot for all jaccard coefficients amongst the three epilepsy ontologies
}
\examples{
\dontrun{
tanimotobaseline <- createTanimotoBaseline(neuroepso, neuroesso, neuroepi, dneuromaxk)
}
}
