% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVennDiagramms.R
\name{drawVenn4}
\alias{drawVenn4}
\title{Create quad Venn Diagramm for overlapping concepts 
between EpSO, ESSO, EPILONT and EPISEM}
\usage{
drawVenn4()
}
\value{
plot object
}
\description{
Create quad Venn Diagramm for overlapping concepts 
between EpSO, ESSO, EPILONT and EPISEM
}
\examples{
\dontrun{
ggplot2::ggsave("venn4.png", plot = drawVenn4(), width=240, height=160, 
  units = "mm", dpi = 300)
}
}
