% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVennDiagramms.R
\name{drawVenn5}
\alias{drawVenn5}
\title{Create quintuple Venn Diagramm for overlapping concepts 
between EpSO, ESSO, EPILONT, EPISEM and FENICS}
\usage{
drawVenn5()
}
\value{
plot object
}
\description{
Create quintuple Venn Diagramm for overlapping concepts 
between EpSO, ESSO, EPILONT, EPISEM and FENICS
}
\examples{
\dontrun{
ggplot2::ggsave("venn5.png", plot = drawVenn5(), width=240, height=160, 
  units = "mm", dpi = 300)
}
}
