% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{drawVennPlot}
\alias{drawVennPlot}
\title{Draw venn diagram with example values}
\usage{
drawVennPlot()
}
\value{
venn diagram
}
\description{
Draw venn diagram with example values
}
\examples{
\dontrun{
drawVennPlot()
}
}
