% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOntStatsIntoEnv.R
\name{genDictListFromRawFreq}
\alias{genDictListFromRawFreq}
\title{Clears object that was loaded from harddrive into a list of terms sorted by frequency}
\usage{
genDictListFromRawFreq(topfreqdictraw)
}
\arguments{
\item{topfreqdictraw}{}
}
\value{
a sorted list of terms
}
\description{
Clears object that was loaded from harddrive into a list of terms sorted by frequency
}
\examples{
\dontrun{
genDictListFromRawFreq(epi)
}
}
