% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoReader.R, R/processATC.R
\name{genDictListFromRawFreq}
\alias{genDictListFromRawFreq}
\title{Clears object that was loaded from harddrive into a list of terms sorted by frequency}
\usage{
genDictListFromRawFreq(topfreqdictraw)

genDictListFromRawFreq(topfreqdictraw)
}
\arguments{
\item{topfreqdictraw}{list with terms from a dictionary sorted by frequency}
}
\value{
a sorted list of terms

a sorted list of terms
}
\description{
Clears object that was loaded from harddrive into a list of terms sorted by frequency

Clears object that was loaded from harddrive into a list of terms sorted by frequency
}
\examples{
\dontrun{
genDictListFromRawFreq(epi)
}
utils::data(rawDrugNamesCoOcEpSO, package="epos")
genDictListFromRawFreq(rawDrugNamesCoOcEpSO)
}
