% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoReader.R
\name{getTermMatrix}
\alias{getTermMatrix}
\title{Receives a sorted hashmap with found entities from a dictionary}
\usage{
getTermMatrix(dictionary, database, collection)
}
\arguments{
\item{dictionary}{Character vector that is the name of a dicitonary having pre-calculated stats. This can be
MeSH, DrugBank, Agrovoc, EpSO, ESSO, or EPILONT}

\item{database}{the name of the MongoDB database to be used}

\item{collection}{the name of the MongoDB collection to be used}
}
\value{
a sorted hashmap containing all found entities from the respective dictionaries with frequencies
}
\description{
Receives a sorted hashmap with found entities from a dictionary
}
\examples{
\dontrun{
mesh <- getTermMatrix("MeSH")
}

}
