% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSimilarities.R
\name{jaccard}
\alias{jaccard}
\title{Calculate jaccard similarity coefficient}
\usage{
jaccard(ainterb, aunionb)
}
\arguments{
\item{ainterb}{integer value with number of intersecting elements between set a and b}

\item{aunionb}{integer value with number of union elements between set a and b}
}
\value{
jac double value with the jaccard similarity coefficient
}
\description{
Calculate jaccard similarity coefficient
}
\examples{
\dontrun{
jaccard(1,3)
}
}
