% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNeuroTable.R
\name{sortTableByRefMatches}
\alias{sortTableByRefMatches}
\title{Sort table by scoring for each row}
\usage{
sortTableByRefMatches(dntk)
}
\arguments{
\item{dntk}{the table returned from writeNeuroTable}
}
\value{
the sorted table
}
\description{
Sort table by scoring for each row
}
\examples{
utils::data(rawDrugNamesCoOcEpSO, package="epos")
utils::data(rawDrugNamesCoOcESSO, package="epos")
utils::data(rawDrugNamesCoOcEPILONT, package="epos")
utils::data(rawDrugNamesCoOcEPISEM, package="epos")
utils::data(rawDrugNamesCoOcFENICS, package="epos")
atchashda <-
readAtcMapIntoHashMapDrugNamesAtcCodes(
  system.file("extdata", "db-atc.map", package = "epos"), "\t")
atchashaa <-
  readAtcMapIntoHashMapAtcCodesAtcNames(
    system.file("extdata", "db-atc.map", package = "epos"), "\t")
atchashsec <-
  readSecondLevelATC(
    system.file("extdata", "atc-secondlevel.map", package = "epos"), "\t")
epso <- genDictListFromRawFreq(rawDrugNamesCoOcEpSO)
neuroepso <- filterNeuroDrugs(epso, atchashda)
esso <- genDictListFromRawFreq(rawDrugNamesCoOcESSO)
neuroesso   <- filterNeuroDrugs(esso, atchashda)
epi <- genDictListFromRawFreq(rawDrugNamesCoOcEPILONT)
neuroepi    <- filterNeuroDrugs(epi, atchashda)
episem <- genDictListFromRawFreq(rawDrugNamesCoOcEPISEM)
neuroepisem <- filterNeuroDrugs(episem, atchashda)
fenics <- genDictListFromRawFreq(rawDrugNamesCoOcFENICS)
neurofenics <- filterNeuroDrugs(fenics, atchashda)
mx <- max(
    c(length(neuroepso), length(neuroesso), length(neuroepi),
      length(neuroepisem), length(neurofenics)))
dneuro <-
  data.frame(EpSO = c(neuroepso, rep("", (mx-length(neuroepso)))),
             ESSO = c(neuroesso, rep("", (mx-length(neuroesso)))),
             EPILONT = c(neuroepi, rep("", (mx-length(neuroepi)))),
             EPISEM = c(neuroepisem, rep("", (mx-length(neuroepisem)))),
             FENICS = c(neurofenics, rep("", (mx-length(neurofenics)))))
dneuromaxk <- TopKLists::calculate.maxK(dneuro, L=5, d=5, v=10)
neurotable <- createNeuroTable(atchashda, atchashsec, dneuromaxk)
sortedNeuroTable <- sortTableByRefMatches(neurotable)
}
